+++ gluecodium/src/test/resources/smoke/listeners/output/android/jni/com_example_smoke_CalculatorListenerImplCppProxy.h
@@ -2,7 +2,7 @@
  *
  */
 #pragma once
-#include "smoke/CalculatorListener.h"
+#include "smoke\CalculatorListener.h"
 #include "CppProxyBase.h"
 #include "JniReference.h"
 namespace gluecodium
@@ -18,7 +18,7 @@ public:
     void on_calculation_result_const( const double ncalculationResult ) const override;
     void on_calculation_result_struct( const ::smoke::CalculatorListener::ResultStruct& ncalculationResult ) override;
     void on_calculation_result_array( const ::std::vector< double >& ncalculationResult ) override;
-    void on_calculation_result_map( const ::smoke::CalculatorListener::NamedCalculationResults& ncalculationResults ) override;
+    void on_calculation_result_map( const ::std::unordered_map< ::std::string, double >& ncalculationResults ) override;
     void on_calculation_result_instance( const ::std::shared_ptr< ::smoke::CalculationResult >& ncalculationResult ) override;
 };
 }